# openapi.yaml
openapi: 3.0.0
info:
  title: SecureSKYE
  description: ''
  version: 1.0.0
servers:
  - url: secureskye-demo.trilliumsecure.com/api
  - url: 192.168.10.220/api
...
paths:
...
  '/users/{id}':
    get:
      tags:
        - user-admin
      operationId: getUser
      description: Get a single user by its ID.
      parameters:
        - name: id
          in: path
          description: Identifier of the user to retrieve.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
      security:
        - webapp_auth:
            - 'user:read'
    patch:
      tags:
        - user-admin
      operationId: updateUser
      description: Update one or more details about a single user.
      parameters:
        - name: id
          in: path
          description: Identifier of the user to update.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Credential'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
      security:
        - webapp_auth:
            - 'user:update'
    delete:
      tags:
        - user-admin
      operationId: deleteUser
      description: Completely delete a user from the system.
      parameters:
        - name: id
          in: path
          description: Identifier of the user to delete.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
      security:
        - webapp_auth:
            - 'user:delete'
...
  '/internal/users':
    get:
      tags:
        - user-admin
      operationId: listUsersInternal
      description: Internal endpoint for listing all existing users.
      parameters:
        - name: name
          in: query
          schema:
            type: string
          description: Optional name to find a single user
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CredentialList'
      security:
        - webapp_auth:
            - 'user:read'
  '/internal/users/{value}':
    get:
      tags:
        - user-admin
      operationId: getUserInternal
      description: Internal endpoint to get a single user by either ID or name.
      parameters:
        - name: value
          in: path
          description: >-
            The value to get a user by. This depends on the `field` query
            parameter.
          required: true
          schema:
            type: string
        - name: field
          in: query
          description: >-
            Define which value is represented in the `value` query parameter,
            effectively choosing how to search a user.
          schema:
            type: string
            enum:
              - id
              - name
            default: id
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Credential'
      security:
        - webapp_auth:
            - 'user:read'
...
  /vehicles:
    get:
      tags:
        - vehicle-admin
      operationId: listVehicles
      description: List all vehicles registered in the system.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VehicleList'
      security:
        - webapp_auth:
            - 'vehicle:read'
    post:
      tags:
        - vehicle-admin
      operationId: createVehicle
      description: Register a new vehicle with the system.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewVehicle'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisteredVehicle'
      security:
        - webapp_auth:
            - 'vehicle:create'
  '/vehicles/{id}':
    get:
      tags:
        - vehicle-admin
      operationId: getVehicle
      description: Get a single specific vehicle.
      parameters:
        - name: id
          in: path
          description: >-
            Identifier of the vehicle to retrieve. This can also be the vehicles
            client ID.
          required: true
          schema:
            type: string
            pattern: ^(?!\s*$).+
        - name: clientId
          in: query
          description: Wherether the provided ID is the direct vehicle ID or its client ID.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
      security:
        - webapp_auth:
            - 'vehicle:read'
    delete:
      tags:
        - vehicle-admin
      operationId: deleteVehicle
      description: Delete a single vehicle from the system.
      parameters:
        - name: id
          in: path
          description: Identifier of the vehicle to delete.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
      security:
        - webapp_auth:
            - 'vehicle:delete'
...
  '/internal/vehicles/{id}':
    get:
      tags:
        - vehicle-admin
      operationId: getVehicleInternal
      description: Internal endpoint for getting information about a single vehicle.
      parameters:
        - name: id
          in: path
          description: >-
            Identifier of the vehicle to get information about. Or client ID of
            the vehicle.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: clientId
          in: query
          description: Wherether the provided ID is the direct vehicle ID or its client ID.
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Vehicle'
      security:
        - webapp_auth:
            - 'vehicle:read'
  '/internal/vehicles/{id}/alert-count':
    put:
      tags:
        - vehicle-admin
      operationId: updateAlertCount
      description: Update vehicle internal alert count by vehicle Id
      parameters:
        - name: id
          in: path
          description: Vehicle Id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleAlertCount'
      responses:
        '200':
          description: OK
      security:
        - webapp_auth:
            - 'vehicle:update'
  '/internal/vehicles/{id}/connection':
    post:
      tags:
        - vehicle-admin
      operationId: saveConnectionStatus
      description: Save vehicle internal connection status by vehicle Id
      parameters:
        - name: id
          in: path
          description: Vehicle Id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VehicleConnection'
      responses:
        '200':
          description: OK
      security:
        - vehicle_auth:
            - 'vehicle-status:update'
  '/internal/vehicles/{id}/connection/{status}':
    put:
      tags:
        - vehicle-admin
      operationId: updateTrackingStatus
      description: Update vehicle internal connection status by vehicle Id
      parameters:
        - name: id
          in: path
          description: Vehicle Id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: status
          in: path
          description: The status of vehicle
          required: true
          schema:
            $ref: '#/components/schemas/VehicleTrackingStatus'
      responses:
        '200':
          description: OK
      security:
        - vehicle_auth:
            - 'vehicle-status:update'
  '/internal/vehicles/{id}/update-count':
    put:
      tags:
        - vehicle-admin
      operationId: updateVehicleUpdateCount
      description: Update vehicle internal update count by vehicle Id
      parameters:
        - name: id
          in: path
          description: Vehicle Id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: updateCount
          required: true
          in: query
          description: Unknow
          schema:
            type: integer
            format: int64
            default: 0
      responses:
        '200':
          description: OK
      security:
        - webapp_auth:
            - 'vehicle:update'
...
  '/vehicles/events':
    post:
      tags:
        - vehicle-message
      operationId: saveEventBatch
      description: >-
        Save a new batch of vehicle events. The batch can contain any amount of
        the supported events. This endpoint can only be called by vehicles
        directly and the vehicle ID is auto-detected by the login.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EventBatch'
      responses:
        '200':
          description: OK
      security:
        - vehicle_auth:
            - 'event:create'
...
  '/vehicles/{id}/events':
    get:
      tags:
        - vehicle-message
      operationId: listRecentEvents
      description: Get all most recent vehicle events.
      parameters:
        - name: id
          in: path
          description: Identifier of the vehicle to get events from.
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: fake
          in: query
          description: Fake data
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          description: Limit the amount of each event type returned.
          schema:
            type: integer
            format: int32
            default: 10
            minimum: 1
            maximum: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventBatch'
      security:
        - webapp_auth:
            - 'event:read'
...
  '/obd2devices/{id}/vehicle/{vehicleId}':
    put:
      tags:
        - obd2device-admin
      operationId: registerAnObd2DeviceToVehicle
      description: Register an obd2 device to vehicle by obd2device Id and vehicle Id
      parameters:
        - name: id
          in: path
          description: OBD-II device ID
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
        - name: vehicleId
          in: path
          description: Vehicle Id
          required: true
          schema:
            $ref: '#/components/schemas/UUID'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Obd2Device'
...
components:
  securitySchemes:
    webapp_auth:
      type: oauth2
      flows:
        password:
          tokenUrl: 'https://secureskye-demo.trilliumsecure.com/uaa/oauth/token'
          scopes:
            'anomaly:read': ''
            'event:read': ''
            'route:create': ''
            'route:read': ''
            'route:update': ''
            'route:delete': ''
            'vehicle:create': ''
            'vehicle:read': ''
            'vehicle:update': ''
            'vehicle:delete': ''
            'vehicle-model:create': ''
            'vehicle-model:read': ''
            'vehicle-model:update': ''
            'vehicle-model:delete': ''
            'vehicle-registry:read': ''
            'vehicle-status:update': ''
            'user:create': ''
            'user:read': ''
            'user:update': ''
            'user:delete': ''
            'user-settings:read': ''
            'user-settings:update': ''
            'authority:read': ''
            'ie:create': ''
            'ie:read': ''
            'message:read': ''
            'ota-images:create': ''
            'ota-images:read': ''
            'ota-images:update': ''
            'ota-images:delete': ''
            'notification:create': ''
            'notification:read': ''
            'notification:delete': ''
            'software-install:create': ''
...
  schemas:
...
    UUID:
      type: string
      pattern: '^[a-z0-9]{8}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{4}-[a-z0-9]{12}$'
      description: A universally unique identifier used as ID for most of the objects.
...
    TwoFactorAuthenticationStatus:
      type: integer
      format: int32
      enum:
      - 0
      - 1
      - 2
      description: |
        The current status of two factor authentication.
        * 0 - Disabled
        * 1 - Enabled
        * 2 - Pending
    Credential:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        version:
          type: integer
          format: int32
        name:
          type: string
          description: The username used for login.
        password:
          type: string
          description: Password of the user. This is usually empty.
        enabled:
          type: boolean
          description: Werether the user account is disabled.
        avatar:
          type: string
          description: >-
            The user's profile picture. On requests this is the Base64 encoded
            data of the image. On responses this is an URL to download the
            image.
        avatarFormat:
          type: string
          description: Format (e.g. PNG) of the image.
        avatarThirdParty:
          type: boolean
          description: >-
            Werether this avatar is from a 3rd party service or provided
            directly by the service.
        firstName:
          type: string
          description: First name of the user.
        lastName:
          type: string
          description: Last (family) name of the user.
        email:
          type: string
          description: Main contact and account eMail address.
        phoneAreaCode:
          type: string
        phoneNumber:
          type: integer
        gender:
          type: boolean
        birthday:
          type: integer
          format: int64
        nationality:
          type: string
        address:
          type: string
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
          description: A list of all roles assigned to the user.
        tfa:
          $ref: '#/components/schemas/TwoFactorAuthenticationStatus'
        additionalInfo:
          type: array
          items:
            $ref: '#/components/schemas/AdditionalInfo'
        attachments:
          type: array
          items:
            $ref: '#/components/schemas/Attachment'
      description: User login credentials and basic information.
    CredentialList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Credential'
          description: The actual data list.
      description: A list of users.
    Role:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Unique name of this role.
        description:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/Credential'
          description: List of all users that this role owns.
        authorities:
          type: array
          items:
            $ref: '#/components/schemas/Authority'
          description: List of all authorities that this role owns.
      description: >-
        A role that bundles several authorities into one entity. This is similar
        to groups.
...
    Authority:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        authority:
          type: string
          description: Name of this authority.
        description:
          type: string
          description: >-
            Short description about the access rights available through this
            authority.
      description: An authority which describes the smallest unit of access right.
...
    AdditionalInfo:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        key:
          type: string
        value:
          type: string
    Attachment:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        fileId:
          type: string
        fileName:
          type: string
        url_image:
          type: string
...
    NewVehicle:
      type: object
      properties:
        modelId:
          $ref: '#/components/schemas/UUID'
        name:
          type: string
          description: Name of the new vehicle.
      description: Request for a new vehicle to be created.
...
    VehicleBodyType:
      type: string
      enum:
        - CUSTOM
        - HATCHBACK
        - SEDAN
        - MUV
        - SUV
        - COUPE
        - CONVERTIBLE
        - WAGON
        - VAN
        - JEEP
...
    Vehicle:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        modelId:
          $ref: '#/components/schemas/UUID'
        modelName:
          type: string
          description: Name of the vehicle model it belongs to.
        name:
          type: string
          description: Name of this vehicle.
        tfa:
          $ref: '#/components/schemas/TwoFactorAuthenticationStatus'
        makerId:
          $ref: '#/components/schemas/UUID'
        makerName:
          type: string
          description: Name of the vehicle maker it belongs to.
        bodyType:
          $ref: '#/components/schemas/VehicleBodyType'
        vin:
          type: string
        color:
          type: string
        imageUrl:
          type: string
        updateCount:
          type: integer
        connection:
          $ref: '#/components/schemas/VehicleConnection'
        alertCount:
          $ref: '#/components/schemas/VehicleAlertCount'
        deleted:
          type: boolean
      description: Representation of a single vehicle.
    VehicleList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Vehicle'
          description: The actual data list.
      description: A list of vehicles.
    VehicleAlertCount:
      type: object
      properties:
        danger:
          type: number
          format: int32
        warning:
          type: number
          format: int32
        info:
          type: number
          format: int32
      description: total count of each alert type of a vehicle
    VehicleConnection:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UUID'
        connected:
          type: boolean
        connectedTimestamp:
          type: number
          format: int64
        disconnectedTimestamp:
          type: number
          format: int64
        lastReceivingTimestamp:
          type: number
          format: int64
        ipAddress:
          type: string
        status:
          $ref: '#/components/schemas/VehicleTrackingStatus'
      description: Connection status and information of a vehicle
    VehicleTrackingStatus:
      type: string
      enum:
      - UNKNOWN
      - STOPPED
      - MOVING
      description: Vehicle moving status
    RegisteredVehicle:
      type: object
      properties:
        vehicle:
          $ref: '#/components/schemas/Vehicle'
        clientId:
          type: string
          description: >-
            The client ID to be used for authenticating requests against the
            service.
      description: A newly registered vehicle.
...
    Obd2Device:
      type: object
      properties:
        id:
          type: string
        family:
          type: string
        kernel:
          type: string
        macAddress:
          type: string
        vehicleId:
          $ref: '#/components/schemas/UUID'
      description: Information of an Obd2 device
...